+++ lib/holiday-stops/src/main/scala/com/gu/salesforce/holiday_stops/SalesforceHolidayStopRequest.scala
@@ -6,7 +6,7 @@ import java.util.UUID
 
 import ai.x.play.json.Jsonx
 import com.gu.holiday_stops.ActionCalculator
-import com.gu.holiday_stops.subscription.CreditCalculator.PartiallyWiredCreditCalculator
+import com.gu.holiday_stops.subscription.Credit.PartiallyWiredCreditCalculator
 import com.gu.holiday_stops.subscription.Subscription
 import com.gu.salesforce.RecordsWrapperCaseClass
 import com.gu.salesforce.SalesforceConstants._
