+++ lib/holiday-stops/src/test/scala/com/gu/holiday_stops/subscription/CreditCalculatorSpec.scala
@@ -36,6 +36,6 @@ class CreditCalculatorSpec extends FlatSpec with Matchers with EitherValues {
     val subscriptionRaw = Source.fromResource(zuoraSubscriptionData).mkString
     val subscription = decode[Subscription](subscriptionRaw).getOrElse(fail(s"Could not decode $zuoraSubscriptionData"))
 
-    CreditCalculator(Fixtures.config)(stopDate, subscription) should equal(Right(expectedCredit))
+    Credit(Fixtures.config)(stopDate, subscription) should equal(Right(expectedCredit))
   }
 }
